third_party/blink/web_tests/external/wpt/css/css-syntax/declarations-trim-whitespace-expected.txt

This is a testharness.js-based test.
PASS --foo-1:bar;
FAIL --foo-2: bar; assert_equals: expected "bar" but got " bar"
FAIL --foo-3:bar ; assert_equals: expected "bar" but got "bar "
FAIL --foo-4: bar ; assert_equals: expected "bar" but got " bar "
FAIL --foo-5: bar !important; assert_equals: expected "bar" but got " bar "
FAIL --foo-6: bar !important ; assert_equals: expected "bar" but got " bar "
PASS --foo-7:bar!important;
PASS --foo-8:bar!important ;
FAIL --foo-9:bar (then ws until end of rule) assert_equals: expected "bar" but got "bar "
Harness: the test ran to completion.
